+++ src/main/java/org/example/dongmoo/section01/Problem06.java
@@ -1,0 +1,20 @@
+package org.example.dongmoo.section01;
+
+import java.io.*;
+
+public class Problem06 {
+    public static String solution(String str) {
+        String answer = "";
+        for (int i = 0; i < str.length(); i++) {
+            if (i == str.indexOf(str.charAt(i))) {
+                answer += str.charAt(i);
+            }
+        }
+        return answer;
+    }
+
+    public static void main(String[] args) throws IOException {
+        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+        System.out.println(solution(br.readLine()));
+    }
+}
