+++ src/main/java/org/example/dongmoo/section01/Problem08.java
@@ -1,28 +1,4 @@
 package org.example.dongmoo.section01;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-
 public class Problem08 {
-    public static String solution(String str) {
-        String answer = "YES";
-        str = str.toLowerCase();
-        int length = str.length();
-
-        for (int i = 0; i < length / 2; i++) {
-            if (str.charAt(i) != str.charAt(length - i - 1)) {
-                return "NO";
-            }
-        }
-        return answer;
-
-
-    }
-
-    public static void main(String[] args) throws IOException {
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        String answer = solution(br.readLine());
-        System.out.println(answer);
-    }
 }
